Transfer state to the board
The state was encapsulated in each square component, however, this
information has to somehow be transferred to the board to keep track of
the game in order to determine who has won and placing alternate X and O
values. It is possible for the board to inquire each square of its state
but that can make code difficult to understand and become complex and
later difficult to refactor.

Instead, it is advisable to lift the state up one level, to the board
component. The Board component can tell each Square what to display.

The handleClick method uses the slice() method to copy the current
squares array on the board state. This is to avoid mutation, which can
make some features more complex to implement. For example, tracking
changes of an immutable object is easier has we have a reference to the
previous and current state, all that is required is to compare the two
to determine whether a change has occured. With a mututated object, we
no longer have a reference to the previous state.

with '#' will be ignored, and an empty message aborts the commit. # #
Date:      Mon May 7 14:16:00 2018 +0100 # # On branch master # Changes
to be committed: #	modified:   src/index.js #